# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Knowledge Melting Pot API
  version: 1.0.0
  description: Versioned API for auditable knowledge extraction and publication.
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
security:
  - sessionCookie: []
paths:
...
  /users:
    get:
      summary: List administrator-managed local accounts
      operationId: listUsers
      responses:
        '200':
          description: Local accounts without password hashes.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
    post:
      summary: Create a local account with a mandatory first-login password change
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: Created account.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /users/{userId}:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    patch:
      summary: Update a local account's display name, activation state, or roles
      operationId: updateUser
      requestBody:
        required: true
        content:
          application/merge-patch+json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated account; role or activation changes revoke existing sessions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /users/{userId}/password-reset:
    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Reset a user's password and require a first-login change
      description: >-
        ADMIN only. The new password is write-only and is never logged. The account must change its
        password at the next login and all of its sessions are revoked. An administrator cannot reset
        their own password here and must use the change-password flow instead.
      operationId: resetUserPassword
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetUserPasswordRequest'
      responses:
        '204':
          description: Password reset; the account must change it at next login and its sessions are revoked.
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
  /skills:
    get:
      summary: List skills (templates and instances) with their latest immutable version
      operationId: listSkills
      parameters:
        - name: kind
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/SkillKind'
        - name: sceneId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Skills with their latest version; manifests are safe resource-only metadata.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Skill'
        '401':
          $ref: '#/components/responses/Problem'
    post:
      summary: Create a global TEMPLATE skill with its first immutable version
      description: ADMIN only. The manifest is validated as resource-only JSON metadata and is never executed.
      operationId: createSkill
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSkillRequest'
      responses:
        '201':
          description: Created template with its first version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /skills/{skillId}:
    parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get a skill with its immutable version history
      operationId: getSkill
      responses:
        '200':
          description: Skill detail and versions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDetail'
        '404':
          $ref: '#/components/responses/Problem'
  /skills/{skillId}/instances:
    parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Fork a scene INSTANCE from a TEMPLATE (OPERATOR/ADMIN)
      description: The instance inherits the template name and its latest version.
      operationId: forkSkillInstance
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInstanceRequest'
      responses:
        '201':
          description: Created scene instance.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Skill'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
  /skills/{skillId}/versions:
    parameters:
      - name: skillId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Append a new immutable version to an INSTANCE skill (OPERATOR/ADMIN)
      operationId: createSkillVersion
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateVersionRequest'
      responses:
        '201':
          description: Appended immutable version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillVersion'
        '400':
          $ref: '#/components/responses/Problem'
        '401':
          $ref: '#/components/responses/Problem'
        '403':
          $ref: '#/components/responses/Problem'
        '404':
          $ref: '#/components/responses/Problem'
        '409':
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: KMP_SESSION
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
  responses:
...
    Problem:
      description: RFC 7807 problem response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    CreateUserRequest:
      type: object
      additionalProperties: false
      required: [username, displayName, initialPassword, roles]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
          pattern: '^[A-Za-z0-9][A-Za-z0-9._-]{2,99}$'
        displayName:
          type: string
          minLength: 1
          maxLength: 200
        initialPassword:
          type: string
          minLength: 12
          maxLength: 128
          writeOnly: true
        roles:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [OPERATOR, PUBLISHER, ADMIN]
    UpdateUserRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 200
        enabled: { type: boolean }
        roles:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            type: string
            enum: [OPERATOR, PUBLISHER, ADMIN]
    ResetUserPasswordRequest:
      type: object
      additionalProperties: false
      required: [newPassword]
      properties:
        newPassword:
          type: string
          minLength: 12
          maxLength: 128
          writeOnly: true
    User:
      type: object
      additionalProperties: false
      required: [id, username, displayName, roles, enabled, mustChangePassword]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        displayName: { type: string }
        roles:
          type: array
          uniqueItems: true
          items:
            type: string
            enum: [OPERATOR, PUBLISHER, ADMIN]
        enabled: { type: boolean }
        mustChangePassword: { type: boolean }
...
    SkillKind:
      type: string
      enum: [TEMPLATE, INSTANCE]
    CreateSkillRequest:
      type: object
      additionalProperties: false
      required: [name, manifest, packageHash]
      properties:
        name: { type: string, minLength: 1, maxLength: 200 }
        description: { type: string, maxLength: 2000 }
        manifest:
          type: string
          description: >-
            Safe resource-only JSON metadata (resources/prompt/schema/files).
            Must declare executionMode RESOURCE_ONLY; scripts, binaries, commands,
            and secrets are rejected and the manifest is never executed.
        packageHash:
          type: string
          pattern: '^[0-9a-f]{64}$'
    CreateInstanceRequest:
      type: object
      additionalProperties: false
      required: [sceneId]
      properties:
        sceneId: { type: string, format: uuid }
    CreateVersionRequest:
      type: object
      additionalProperties: false
      required: [manifest, packageHash]
      properties:
        manifest:
          type: string
          description: Safe resource-only JSON metadata; never executed.
        packageHash:
          type: string
          pattern: '^[0-9a-f]{64}$'
    Skill:
      type: object
      additionalProperties: false
      required: [id, name, kind, description, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { $ref: '#/components/schemas/SkillKind' }
        description: { type: string }
        sceneId: { type: [string, 'null'], format: uuid }
        sourceSkillId: { type: [string, 'null'], format: uuid }
        sourceSkillVersionId: { type: [string, 'null'], format: uuid }
        version: { type: [integer, 'null'], minimum: 1 }
        packageHash: { type: [string, 'null'] }
        manifestJson: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    SkillVersion:
      type: object
      additionalProperties: false
      required: [id, skillId, version, manifestJson, packageHash, createdBy, createdAt]
      properties:
        id: { type: string, format: uuid }
        skillId: { type: string, format: uuid }
        version: { type: integer, minimum: 1 }
        manifestJson: { type: string }
        packageHash: { type: string }
        createdBy: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
    SkillDetail:
      type: object
      additionalProperties: false
      required: [id, name, kind, description, createdAt, versions]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        kind: { $ref: '#/components/schemas/SkillKind' }
        description: { type: string }
        sceneId: { type: [string, 'null'], format: uuid }
        sourceSkillId: { type: [string, 'null'], format: uuid }
        sourceSkillVersionId: { type: [string, 'null'], format: uuid }
        createdAt: { type: string, format: date-time }
        versions:
          type: array
          items:
            $ref: '#/components/schemas/SkillVersion'
...
    Problem:
      type: object
      required: [type, title, status, code, traceId]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer, minimum: 400, maximum: 599 }
        detail: { type: string }
        instance: { type: string, format: uri-reference }
        code: { type: string }
        traceId: { type: string }
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string }
              message: { type: string }